Setting up the battlefield
I've jus set up a battlefield using the tile prefab. A prefab is similar to a class as we can create a lot of duplicate objects with the same attributes and scripts.
I've also created the tile.cs script to randomize the tile appearance. I'm hoping to use this funtionality in future where tiles that look differently will have different moving costs